adjusted event modal
